+++ backend/src/baserow/contrib/database/models.py
@@ -2,13 +2,15 @@ from baserow.core.models import Application
 
 from .table.models import Table
 from .views.models import View, GridView
-from .fields.models import Field, TextField, NumberField, BooleanField
+from .fields.models import (
+    Field, TextField, NumberField, LongTextField, BooleanField, DateField
+)
 
 __all__ = [
     'Database',
     'Table',
     'View', 'GridView',
-    'Field', 'TextField', 'NumberField', 'BooleanField',
+    'Field', 'TextField', 'NumberField', 'LongTextField', 'BooleanField', 'DateField',
 ]